Switch FTRL to `float`
