+++ c/models/py_ftrl.h
@@ -25,7 +25,7 @@
 
 namespace py {
 
-template <typename T> 
+template <typename T>
 using dtftptr = std::unique_ptr<dt::Ftrl<T>>;
 
 enum class RegType : uint8_t {
@@ -38,7 +38,7 @@ enum class RegType : uint8_t {
 
 class Ftrl : public PyObject {
   private:
-    std::vector<dtftptr<double>>* dtft;
+    std::vector<dtftptr<float>>* dtft;
     DataTable* feature_names;
     py::olist labels;
     RegType reg_type;
@@ -55,9 +55,10 @@ class Ftrl : public PyObject {
     };
 
     // Initializers and destructor
+    void m__initT__(PKArgs&);
     void m__init__(PKArgs&);
     void m__dealloc__();
-    void init_dtft(dt::FtrlParams<double>);
+    void init_dtft(dt::FtrlParams<float>);
 
     // Pickling support
     oobj m__getstate__(const PKArgs&);
@@ -107,7 +108,7 @@ class Ftrl : public PyObject {
     static void normalize_rows(DataTable*);
 
     // Helper functions
-    static void normalize_fi(RealColumn<double>*);
+    static void normalize_fi(RealColumn<float>*);
 };
 
 } // namespace py
